Work around incorrect element detection on iOS 6 in transitioning/scrolling layers (fixes #57)
iOS 6.0 (and potentially beyond) has a bug where event.target gives an
incorrect element if that element is within a parent that is currently
being scrolled or CSS transitioned.  This works around the problem by
performing manual detection in these circumstanced by using
document.elementFromPoint with coordinates derived from the event,
which works if preceded by an event.target.