iPhone 5 users not able to tap
Commenting these lines seems to fix an issue we had with our iPhone5 customers where they are not able to tap/click on buttons/textboxes but they can submit using keypad Go button.
https://github.com/ftlabs/fastclick/blob/master/lib/fastclick.js#L388-L391

It might have some adverse behaviour, but not sure what it would be.

Also to note that it has surprisingly came up today when many of our customers reported it when the same code is in production from long time. Might be that Apple/webkit has released an update.