Wrap whole lib in IIFE to circumvent problems with eval-based devtools
It fixes a problem when fastclick is used with a bundler like browserify or webpack which
use eval() for source urls.

According to the documentation on eval() in strict mode "strict" code can only access
variables defined by eval() in the same scope. Thus accessing FastClick or deviceIsAndroid
inside of a function (which is preceded by 'use strict') triggers a ReferenceError.

#270