Clarify Twitter Bootstrap dropdown compatibility
"Use case 2" in the documentation describes a compatibility issue with
Twitter Bootstrap 2.2.2.  After reading this, I was unsure about whether
the issue only applied to version 2.2.2 or whether this workaround was
still required.  After some investigation, it appears that the
compatibility issue was introduced in twbs/bootstrap@c9cef741 (included
in `v2.1.0` and later) and was resolved in twbs/bootstrap@b5ad5068
(included in `v2.3.0` and later, including `v3.0.0`) as a result of
twbs/bootstrap#6488.

To avoid causing confusion for future users who may not know about the
state of this issue, this commit clarifies which versions of Twitter
Bootstrap are affected.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>