Only blur before click on Android
Calling blur before dispatching the click event causes the blur event to
fire before the click event, rather than after as would be the case with
a browser click event.  This can cause issues in other code such as
twitter/typeahead.js#792.

This commit limits the unexpected event order to Android platforms,
where it is required for proper focus behavior (#24).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>